Some fixes

+++ telegramBot.py
@@ -1,7 +1,7 @@
 import logging
 import telegram
 from enum import Enum
-from datetime import datetime
+from datetime import datetime, timedelta
 from dataclasses import dataclass
 from typing import Dict, List, Tuple, Type
 from telegram import InlineKeyboardMarkup, InlineKeyboardButton, Update
@@ -60,7 +60,7 @@ class RespondData:
     UserId: int
     ChatId: int
     MessageId: int
-    StartingTime: datetime
+    StartingTime: str
 
 # Initialize dictionaries to store data
 userData = {}
@@ -73,13 +73,21 @@ async def create(update: Update, context: ContextTypes.DEFAULT_TYPE):
     user_first_name = update.effective_user.first_name
     user_second_name = update.effective_user.last_name
     chat_id = update.effective_chat.id
-    
+
     # Check if the user is an admin in the group or supergroup
     chat_administrators = await context.bot.get_chat_administrators(chat_id)
     is_admin = any(admin.user.id == user_id for admin in chat_administrators)
 
     if not is_admin:
         return
+    if update.message.chat.type == 'private':
+        return
+
+    last_event_data = LastEvent.get(chat_id)
+    if last_event_data:
+        await try_delete_message(update._bot, chat_id=last_event_data.ChatId, message_id=last_event_data.MessageId)
+        LastEvent.pop(chat_id)
+
 
     # Ask the user something in private messages with buttons
     keyboard = [
@@ -96,7 +104,7 @@ async def create(update: Update, context: ContextTypes.DEFAULT_TYPE):
     # Store the state that the user is expected to answer
     user_data = UserData(user_id, chat_id, None, user_first_name, user_second_name)
     userData[user_id] = user_data
-    
+
     await try_delete_message(update._bot, update.message.chat_id, update.message.message_id)
 
 # Function to handle the /edit command
@@ -110,17 +118,19 @@ async def edit(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
     if not is_admin:
         return
+    if update.message.chat.type == 'private':
+        return
     
-    # Check if there is a last event to edit
-    if chat_id not in LastEvent or not LastEvent[chat_id]:
-        context.bot.send_message(chat_id, "Немає івентів для зміни")
+    last_event_data = LastEvent.get(chat_id)
+    if last_event_data is None:
+        await try_delete_message(update._bot, update.message.chat_id, update.message.message_id)
         return
 
     # Ask the user what to edit
     keyboard = [
         [InlineKeyboardButton("Заголовок", callback_data="Header"),
          InlineKeyboardButton("Опис", callback_data="Description")],
-        [InlineKeyboardButton("Данні карти", callback_data="Card Data")],
+        [InlineKeyboardButton("Данні карти", callback_data="Card data")],
         [InlineKeyboardButton("Завершити зміни", callback_data="End Edit")]
     ]
     reply_markup = InlineKeyboardMarkup(keyboard)
@@ -131,7 +141,7 @@ async def edit(update: Update, context: ContextTypes.DEFAULT_TYPE):
     user_data = UserData(user_id, chat_id, LastEvent[chat_id].MessageId)
     user_data.CurrentMod = "None"  # Set the mode to "None" initially for editing
     userData[user_id] = user_data
-    
+
     await try_delete_message(update._bot, update.message.chat_id, update.message.message_id)
 
 # Function to handle the /delete command
@@ -145,10 +155,18 @@ async def delete(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
     if not is_admin:
         return
+    if update.message.chat.type == 'private':
+        return
+    
+    last_event_data = LastEvent.get(chat_id)
+    if last_event_data is None:
+        await try_delete_message(update._bot, update.message.chat_id, update.message.message_id)
+        return
 
     last_event_data = LastEvent[chat_id]
     await context.bot.delete_message(chat_id=last_event_data.ChatId, message_id=last_event_data.MessageId)
-            
+    LastEvent.pop(chat_id)
+
     await try_delete_message(update._bot, update.message.chat_id, update.message.message_id)
 
 # Function to handle messages
@@ -164,11 +182,6 @@ async def handle_update(update: Update, context: ContextTypes.DEFAULT_TYPE):
 
     if update.message:
         message_text = update.message.text
-
-        if user_data.IsEditing:
-            await context.bot.send_message(chat_id=user_id, text="You are currently editing, please finish the previous action.")
-            return
-            
         if update.message.chat.type == 'private':
             if user_id not in userData:
                 await context.bot.send_message(chat_id=chat_id, text="Вибачаюсь, немає подій, які ви зараз налаштовуєте.")
@@ -201,15 +214,15 @@ async def handle_update(update: Update, context: ContextTypes.DEFAULT_TYPE):
 async def handle_callback_query(update: Update, context: ContextTypes.DEFAULT_TYPE):
     user_id = update.effective_user.id
     chat_id = update.effective_chat.id
-    
+
     user_response = update.callback_query.data
-    
+
     if update.callback_query.message.chat.type == Filters.private.name:
         user_data: Type[UserData] = userData[user_id]
 
         if user_data is None:
             return
-        
+
         if user_response == "Header":
             await context.bot.send_message(chat_id=user_id, text="Будь ласка, введіть текст Заголовку:")
             user_data.CurrentMod = CurrentMode.Header
@@ -220,54 +233,56 @@ async def handle_callback_query(update: Update, context: ContextTypes.DEFAULT_TY
             await context.bot.send_message(chat_id=user_id, text="Будь ласка, введіть номер карти:")
             user_data.CurrentMod = CurrentMode.Card
         elif user_response == "Yes" or user_response == "End Edit":
-            
-            
-            
+
             if(user_response == "Yes"):
                 header_text = user_data.Header if user_data.Header is not None else ""
                 description_text = user_data.Description if user_data.Description is not None else ""
                 card_data = user_data.CardData if user_data.CardData is not None else None
                 message_text = await formating_method(header_text, description_text)
                 sent_message = await context.bot.send_message(user_data.ChatId, text=message_text, reply_markup=plus_minus_markup, parse_mode='HTML')
+                last_event = LastEvent.get(user_data.ChatId)
+                if last_event is not None:
+                    await try_delete_message(context.bot, last_event.ChatId, last_event.MessageId)
                 last_event_data = LastEventData(user_data.ChatId, sent_message.message_id, header_text, description_text, card_data)
-            
+
             elif(user_response == "End Edit"):
-                last_event = LastEvent[user_data.ChatId]
+                last_event : Type[LastEventData] = LastEvent[user_data.ChatId]
                 header_text = user_data.Header if user_data.Header is not None else last_event.Header
                 description_text = user_data.Description if user_data.Description is not None else last_event.Description
-                card_data = user_data.CardData if user_data.CardData is not None else None
+                card_data = user_data.CardData if user_data.CardData is not None else last_event.CardData
                 message_text = await format_for_editing(user_data, header_text, description_text, card_data, response_text = None)
                 edited_message = await context.bot.edit_message_text(message_text, last_event.ChatId, last_event.MessageId, reply_markup=plus_minus_markup, parse_mode='HTML')
                 last_event_data = LastEventData(user_data.ChatId, edited_message.message_id, header_text, description_text, card_data)
-            
+
             LastEvent[user_data.ChatId] = last_event_data
 
             user_data.CurrentMod = CurrentMode.NoneMode
             user_data.IsEditing = False
-            
+
             await try_delete_message(update._bot, update.effective_chat.id, update.effective_message.id)
         elif user_response == "No":
             await context.bot.send_message(chat_id=chat_id, text="Ви відмінили створення події")
+            await try_delete_message(update._bot, update.effective_chat.id, update.effective_message.id)
             userData.pop(user_id)
             return
-        
+
     elif update.callback_query.message.chat.type == Filters.group.name or update.callback_query.message.chat.type == Filters.supergroup.name:
         query = update.callback_query
         user_response = query.data
-        
+
         some_user_data = UserData(
         user_id = query.from_user.id,
         chat_id = query.message.chat.id,
         message_id= query.message.message_id,
         user_first_name= query.from_user.first_name,
         user_second_name= query.from_user.last_name)
-        
+
         response_text = ''
         if user_response == '+':
             response_text = 'yes'
         elif user_response == '-':
             response_text = 'no'
-        
+
         last_event_data: Type[LastEventData] = LastEvent[query.message.chat.id]
         try:
             await context.bot.edit_message_text(
@@ -281,7 +296,7 @@ async def handle_callback_query(update: Update, context: ContextTypes.DEFAULT_TY
 
 # Function to handle first formating
 async def formating_method(header_text, description_text):
-    
+
     separator_line = "──────────────────────────"
 
     if description_text is not None:
@@ -291,29 +306,32 @@ async def formating_method(header_text, description_text):
 
 # Function to handle formating
 async def format_for_editing(user_data: UserData, header_text: str, description_text: str, card_data: CardData, response_text: str) -> str:
-    
+
     user_mention = f"<a href=\"tg://user?id={user_data.UserId}\">{user_data.UserFirstName} {user_data.UserSecondName if user_data.UserSecondName else ''}</a>"
 
     current_time = datetime.now()
-    formatted_date = current_time.strftime("%d %B %H:%M")
-    formatted_user_response = f"{user_mention} <i>{formatted_date}</i>"
+    gmt_offset = timedelta(hours=3) 
+
+    # Add the time difference to the current time to get the GMT time
+    gmt_time = current_time + gmt_offset
+    formatted_date_gmt = gmt_time.strftime("%d %B %H:%M")
+    formatted_user_response = f"{user_mention} <i>{formatted_date_gmt}</i>"
 
     tuple_key = (user_data.ChatId, user_data.MessageId)
 
     if response_text is not None:
         if tuple_key not in RespondsOnEvent:
-            RespondsOnEvent[tuple_key] = RespondData([], user_data.UserId, user_data.ChatId, user_data.MessageId, datetime.now())
+            RespondsOnEvent[tuple_key] = RespondData([], user_data.UserId, user_data.ChatId, user_data.MessageId, formatted_date_gmt)
 
         respond = RespondsOnEvent[tuple_key]
 
         if respond.Responds and any(response.startswith(user_mention) for response in respond.Responds) and response_text.lower() == "no":
             respond.Responds = [response for response in respond.Responds if not response.startswith(user_mention) and not response.startswith(f"<s>{user_mention}")]
-            formatted_user_response = f"<s>{user_mention} <i>{formatted_date}</i></s>"
+            formatted_user_response = f"<s>{user_mention} <i>{formatted_date_gmt}</i></s>"
             respond.Responds.append(formatted_user_response)
         elif response_text and response_text.lower() == "yes":
             respond.Responds = [response for response in respond.Responds if not response.startswith(user_mention) and not response.startswith(f"<s>{user_mention}")]
             respond.Responds.append(formatted_user_response)
-            print("respond text" + formatted_user_response)
 
     final_text_list_responses = ""
     if tuple_key in RespondsOnEvent and RespondsOnEvent[tuple_key].Responds:
@@ -323,7 +341,7 @@ async def format_for_editing(user_data: UserData, header_text: str, description_
             counter += 1
 
     separator_line = "──────────────────────────"
-    starting_time_formatted = RespondsOnEvent[tuple_key].StartingTime.strftime("%d %B %H:%M")
+    starting_time_formatted = RespondsOnEvent[tuple_key].StartingTime
 
     if description_text and card_data and card_data.CardNumber and card_data.MoneyAmount:
         return f"<b>{header_text}</b>\n{separator_line}\n{description_text}\n{separator_line}\n{final_text_list_responses}{separator_line}\nВсього: <strong>{len(RespondsOnEvent[tuple_key].Responds)}</strong> людей\nПодію створено - {starting_time_formatted}\n[Номер карти: <code>{card_data.CardNumber}</code> Сумма: <code>{card_data.MoneyAmount}</code>₴]"
@@ -340,7 +358,6 @@ async def try_delete_message(bot : telegram.Bot, chat_id, message_id):
         await bot.delete_message(chat_id=chat_id, message_id=message_id)
     except Exception as ex:
         logging.error(f"Error deleting message: {ex}")
-
 
 if __name__ == '__main__':
     # Set up logging
